+++ src/Server/ServerCore/Systems/WorldManagement/WorldManagementSystem.cs
@@ -53,7 +53,8 @@ public sealed class WorldManagementSystem : ISystem
     public ISet<EventId> EventIdsOfInterest { get; } = new HashSet<EventId>
     {
         EventId.Server_WorldManagement_WorldSegmentLoaded,
-        EventId.Core_WorldManagement_EntityLeaveWorldSegment
+        EventId.Core_WorldManagement_EntityLeaveWorldSegment,
+        EventId.Server_WorldManagement_ResyncPositionedEntity
     };
 
     public int WorkloadEstimate => 80;